aimage/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
djangorestframework==3.14.0
et-xmlfile==1.1.0
flake8==6.0.0
gunicorn==20.1.0
idna==3.4
mccabe==0.7.0
numpy==1.24.1
openai==0.25.0
openpyxl==3.0.10
pandas==1.5.2
pandas-stubs==1.5.2.221213
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
sentry-sdk==1.12.1
six==1.16.0
sqlparse==0.4.3
tqdm==4.64.1
types-pytz==2022.7.0.0
typing-extensions==4.4.0
urllib3==1.26.13
